pyproject.toml
[tool.poetry]
name = "gcs"
version = "0.1.0"
description = ""
authors = ["Oliver Bonham-Carter <obonhamcarter@allegheny.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.7.0"
rich = "^13.3.1"
black = "^23.1.0"
biopython = "^1.81"

[tool.poetry.scripts]
gcs = "gcs.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
